Query imported longidents from the type environment instead of reading .cmi
The code is complex (see the long implementation comment) because the
paths accepted by ppx_import are not necessarily valid paths for the
OCaml type-checker: in `M.S.t`, `S` may be the name of
a sub-module-type instead of a sub-module (and `t` a component of `S`,
as expected). Consider the following example:

`stuff.ml:`

    module type S = sig
      type t = ...
    end

`use.ml:`

    [%import Stuff.S.t]

This patch passes the ppx_import testsuite.