query imported longidents from the type environment instead of reading .cmi
This first patch does not pass the testsuite, because not all
longidents used by ppx_import are actually valid (type-correct)
identifiers. The testsuite contains this example:

stuff.ml:

    module type S = sig
      type t = ...
    end

use.ml:

    [%import Stuff.S.t]

Querying a member of a type signature in this way is not valid in
OCaml, although it makes sense. To support this, we will have to be
more elaborate, and mix type-environment lookup with manual splitting
of the path and exploration of the returned signature.