be more precise about the error handling when calling the typechecker
While the previous commit passes the testsuite just fine, its behavior
is wrong in the cases where errors happen during the import
lookup. For example, if a .cmi file is found to be invalid during the
lookup process, it will ignore this error and eventually give up
finding the imported definition with a generic error message.